Specify order clause in front so that it's actually used.
This fixes a randomly failing test that depends on sort order of
projects.

Remember that in Golang, chained method calls are evaluated
right-to-left. It makes sense that Order() should be on the left of any
Find() call.

To figure out what Gorm is doing, use Debug(): e.g.
db.Debug().Find(...).