Add endpoint to use a Let's Encrypt cert on a domain.
Rough guide to Let's Encrypt process:

1. Register a randomly generated private key with Let's Encrypt (this
identifies our per-domain Let's Encrypt "account").
2. Ask Let's Encrypt what challenge(s) we need to complete to verify
that we own the domain.
3. Get the HTTP challenge details - this requires us to return a
particular response on a particular path.
4. We save the HTTP challenge details in the database - a separate
endpoint will be responsible for serving the challenge response to Let's
Encrypt.
5. Tell Let's Encrypt that we are ready for them to verify our response
to the HTTP challenge.
6. After Let's Encrypt has verified us, generate a random private key
for the certificate and use it for a CSR (certificate signing request).
7. Send CSR to Let's Encrypt - it will return us a certificate, at last.
8. Upload cert (and its private key) to S3.

See https://tools.ietf.org/html/draft-ietf-acme-acme-02 for the ACME
spec (which is what Let's Encrypt implements).