Add watermark (boolean) column to projects.
We'll use this to indicate whether to display a watermark on HTML pages
of the project. We should actually only do this for free projects, but
since we don't have the concept of "free" or "paid" yet, this will allow
us to disable the watermark for certain sites (e.g. the PubStorm
homepage).