Add a watermark to HTML files.
This is only done for projects with the watermark column set to true.

We inject a <script> tag at the end of HTML files that renders a link to
the PubStorm homepage.

The current implementation will slurp in the file contents into program
memory, which should be improved with a streaming version.

For now, we don't modify files > 5MB in size, as a workaround.